requirements.txt
openenv-core
gymnasium
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
pyyaml==6.0.1
openai==1.13.3
httpx==0.27.0
python-multipart==0.0.9
jinja2==3.1.3
python-dotenv==1.0.1
numpy
